test(pts controller): split create and update test suits
